Changed the button on manage_sensor_in_location.html to toggle button

+++ Strata/views.py
@@ -697,13 +697,24 @@ def push_mail(mail_subject="", mail_html_content=""):
 
 
 def start_save_sensor(request, sensor_id, template_name='Convergence/manage_sensor_in_location.html'):
-    task = Task.objects.filter(task_name='Strata.views.run_back_save', task_params="[[" + str(sensor_id) + "], {}]",
+    data={}
+    if request.is_ajax():
+        try:
+            task = Task.objects.filter(task_name='Strata.views.run_back_save', task_params="[[" + str(sensor_id) + "], {}]",
                                locked_at__isnull=True)
-    if task:
-        task.delete()
-    else:
-        run_back_save(sensor_id, repeat=5)
-    sensors = Strata_sensor.objects.all()
+            if task:
+                task.delete()
+            else:
+                run_back_save(sensor_id, repeat=5)
+            sensors = Strata_sensor.objects.all()
+            data['result'] = "success"
+        except:
+            data['result'] = "error"
+            pass
+        return JsonResponse(data)
+
+    data['result']="Not Ajax"
+    return JsonResponse(data)
     #
     # data = {}
     # prepared_data = []
@@ -720,7 +731,7 @@ def start_save_sensor(request, sensor_id, template_name='Convergence/manage_sens
     #                           })
     #
     # data['result'] = prepared_data
-    return redirect(request.META.get('HTTP_REFERER'))
+    # return redirect(request.META.get('HTTP_REFERER'))
 
 
 @login_required
